Before twiking some settings and shit
